Rename PTR to PMR
The market reporting protocol, formerly known as the trade reporting
protocol, will be used to report both entered orders and occurred
trades.
