+++ parity-net/src/main/java/org/jvirtanen/parity/net/pmr/PMRException.java
@@ -1,0 +1,11 @@
+package org.jvirtanen.parity.net.pmr;
+
+import org.jvirtanen.parity.net.ProtocolException;
+
+public class PMRException extends ProtocolException {
+
+    public PMRException(String message) {
+        super(message);
+    }
+
+}
