+++ parity-net/src/main/java/org/jvirtanen/parity/net/pmr/PMRParser.java
@@ -1,6 +1,6 @@
-package org.jvirtanen.parity.net.ptr;
+package org.jvirtanen.parity.net.pmr;
 
-import static org.jvirtanen.parity.net.ptr.PTR.*;
+import static org.jvirtanen.parity.net.pmr.PMR.*;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
@@ -9,13 +9,13 @@ import org.jvirtanen.nassau.MessageListener;
 /**
  * A parser for inbound messages.
  */
-public class PTRParser implements MessageListener {
+public class PMRParser implements MessageListener {
 
     private Trade trade;
 
-    private PTRListener listener;
+    private PMRListener listener;
 
-    public PTRParser(PTRListener listener) {
+    public PMRParser(PMRListener listener) {
         this.trade = new Trade();
 
         this.listener = listener;
@@ -30,7 +30,7 @@ public class PTRParser implements MessageListener {
             trade(buffer);
             break;
         default:
-            throw new PTRException("Unknown message type: " + (char)messageType);
+            throw new PMRException("Unknown message type: " + (char)messageType);
         }
     }
