+++ parity-net/src/main/java/org/jvirtanen/parity/net/pmr/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * PMR is Parity's native market reporting protocol.
+ */
+package org.jvirtanen.parity.net.pmr;
