--- parity-net/src/main/java/org/jvirtanen/parity/net/ptr/PTRException.java
@@ -1,11 +1,0 @@
-package org.jvirtanen.parity.net.ptr;
-
-import org.jvirtanen.parity.net.ProtocolException;
-
-public class PTRException extends ProtocolException {
-
-    public PTRException(String message) {
-        super(message);
-    }
-
-}
