--- parity-net/src/main/java/org/jvirtanen/parity/net/ptr/package-info.java
@@ -1,4 +1,0 @@
-/**
- * PTR is Parity's native trade reporting protocol.
- */
-package org.jvirtanen.parity.net.ptr;
